requirements.txt
# local package install
-e .

# third party packages
pyyaml
dvc
pandas
dvc[gdrive]
sklearn
pandas
pytest
tox
flake8
flask
gunicorn
mlflow